Module actiserve::routes::nodeinfo

Support for providing nodeinfo on /nodeinfo/2.0

The schema for the reponse format can be found here: http://nodeinfo.diaspora.software/ns/schema/2.0#

Structs

NodeInfo

NodeInfo schema version 2.0

Services

The third party sites this server can connect to via their application API.

Software

Metadata about server software in use.

UsageStats

Usage statistics for this server

UserStats

Statistics about the users of this server

Enums

InboundService

The third party sites this server can retrieve messages from for combined display with regular traffic.

OutboundService

Protocol

Protocols that can be supported on this server.

Constants

NODE_INFO_SCHEMA

Functions

get